학과 설정 완료

+++ src/main/java/com/green/muziuniv_be_user/application/department/model/DepartmentPatchReq.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 
 public class DepartmentPatchReq {
     private String status = "0";
-    private int deptId;
+    private Long deptId;
 
 
-    public DepartmentPatchReq(int deptId){
+    public DepartmentPatchReq(Long deptId){
         this.deptId = deptId;
     }
 }
